fixed merge errors
