+++ tetris/Cargo.toml
@@ -10,4 +10,3 @@ glium = "0.32.1"
 rand = "0.8.5"
 image = "0.24"
 rusttype = { version = "0.9.0", features = ["gpu_cache"] }
-arrayvec = { version = "0.4", default-features = false }
